chore: api refactoring vol 2
