+++ src/api/workspace/mod.rs
@@ -1,4 +1,3 @@
-pub mod config;
 pub mod create;
 pub mod enter;
 pub mod remove;
