+++ src/api/workspace/remove.rs
@@ -5,10 +5,9 @@ use bollard::{
 };
 
 use crate::{
-    api::WorkspaceApi,
+    api::{self, WorkspaceApi},
     model::{types::AnyError, volume::CACHE_ROLE},
     util::{
-        crypt,
         labels::{Labels, ROLE},
         ssh,
     },
@@ -41,7 +40,7 @@ impl<'a> WorkspaceApi<'a> {
             for v in volumes {
                 match v {
                     Volume { ref name, .. }
-                        if name == ssh::VOLUME_NAME || name == crypt::VOLUME_NAME =>
+                        if name == ssh::VOLUME_NAME || name == api::crypt::VOLUME_NAME =>
                     {
                         continue;
                     }
