+++ src/cmd/config/mod.rs
@@ -1,0 +1,3 @@
+pub mod edit;
+pub mod show;
+pub mod template;
