+++ src/cmd/mod.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod init;
 pub mod list;
 pub mod new;
