+++ src/util/mod.rs
@@ -1,5 +1,4 @@
 pub mod backend;
-pub mod crypt;
 pub mod git;
 pub mod id;
 pub mod labels;
